+++ src/fr/maxlego08/menu/api/requirement/Action.java
@@ -1,0 +1,14 @@
+package fr.maxlego08.menu.api.requirement;
+
+import org.bukkit.entity.Player;
+
+public interface Action {
+
+    /**
+     * Allows to execute the action
+     *
+     * @param player Who execute the action
+     */
+    void execute(Player player);
+
+}